refactor(population): split HISCALE observation helpers
